more C++ work
