+++ libs/core/pxt.h
@@ -78,11 +78,15 @@ void set_usb_strings(const char *uf2_info);
 
 namespace pins {
 class CodalSPIProxy;
-class CodalSerialProxy;
 } // namespace pins
 
 typedef pins::CodalSPIProxy* SPIDevice;
-typedef pins::CodalSerialProxy* SerialDevice;
+
+namespace serial {
+class CodalSerialDeviceProxy;
+}
+
+typedef serial::CodalSerialDeviceProxy* SerialDevice;
 
 namespace jacdac {
 class JDProxyDriver;
